Moved CkMarshallMsg to its own header
