+++ src/ck-core/charm++.h
@@ -149,14 +149,7 @@ public:
 	inline const CkGroupID *getGroupDepPtr() const { return &(depGroupIDs[0]); }
 };
 
-#include "CkMarshall.decl.h"
-//This is the message type marshalled parameters get packed into:
-class CkMarshallMsg : public CMessage_CkMarshallMsg {
-public: 
-	char *msgBuf;
-};
-
-
+#include "ckmarshall.h"
 
 //A queue-of-messages, like CkMsgQ<CkReductionMsg>
 template <class MSG>
@@ -1232,6 +1225,7 @@ if(CpvAccess(networkProgressCount) >=  p)  \
 #include "readonly.h"
 #include "ckarray.h"
 #include "ckstream.h"
+#include "ckfutures.h"
 #include "waitqd.h"
 #include "sdag.h"
 #include "ckcheckpoint.h"
@@ -1313,45 +1307,6 @@ int CkRegisterEp(const std::string& name, CkCallFnPtr call, int msgIdx, int char
 
   return CkRegisterEpTemplated(combined.c_str(), call, msgIdx, chareIdx, ck_ep_flags);
 }
-
-CkMarshallMsg *CkAllocateMarshallMsgNoninline(int size,const CkEntryOptions *opts);
-inline CkMarshallMsg *CkAllocateMarshallMsg(int size,const CkEntryOptions *opts=NULL)
-{
-	if (opts==NULL) {
-	  CkMarshallMsg *newMemory = new (size,0)CkMarshallMsg;
-	  setMemoryTypeMessage(UsrToEnv(newMemory));
-	  return newMemory;
-	}
-	else return CkAllocateMarshallMsgNoninline(size,opts);
-}
-
-
-
-
-
-
-
-template <typename T> 
-inline T *CkAllocateMarshallMsgT(int size,const CkEntryOptions *opts) 
-{ 
-  int priobits = 0; 
-  if (opts!=NULL) priobits = opts->getPriorityBits(); 
-  //Allocate the message 
-  T *m=new (size,priobits)T; 
-  //Copy the user's priority data into the message 
-  envelope *env=UsrToEnv(m); 
-  setMemoryTypeMessage(env); 
-  if (opts!=NULL) { 
-    CmiMemcpy(env->getPrioPtr(),opts->getPriorityPtr(),env->getPrioBytes()); 
-    //Set the message's queueing type 
-    env->setQueueing((unsigned char)opts->getQueueing()); 
-  } 
-  return m; 
-} 
-
-
-
-
 
 /************************** Debugging Utilities **************/
