+++ src/ck-core/waitqd.h
@@ -2,7 +2,6 @@
 #define _WAITQD_H
 
 #include "charm++.h"
-#include "ckfutures.h"
 #include "waitqd.decl.h"
 
 extern "C" void CkWaitQD(void);
